Separate functionality into different classes
