+++ data/joe.txt
@@ -1,3 +1,4 @@
-[T][0] Make the bed
+[T][1] Make the bed
 [D][0] Finish Homework (by: Aug 28 2025)
 [E][0] Play Basketball (from: Aug 29 2025 to: Aug 29 2025)
+[T][0] 4 Worksheets
